+++ libraries/ui/src/main/java/androidx/media3/ui/LegacyPlayerControlView.java
@@ -54,7 +54,6 @@ import androidx.media3.common.C;
 import androidx.media3.common.MediaLibraryInfo;
 import androidx.media3.common.Player;
 import androidx.media3.common.Player.Events;
-import androidx.media3.common.Player.State;
 import androidx.media3.common.Timeline;
 import androidx.media3.common.util.Assertions;
 import androidx.media3.common.util.RepeatModeUtil;
@@ -840,22 +839,22 @@ public class LegacyPlayerControlView extends FrameLayout {
     }
     boolean requestPlayPauseFocus = false;
     boolean requestPlayPauseAccessibilityFocus = false;
-    boolean shouldShowPauseButton = shouldShowPauseButton();
+    boolean shouldShowPlayButton = Util.shouldShowPlayButton(player);
     if (playButton != null) {
-      requestPlayPauseFocus |= shouldShowPauseButton && playButton.isFocused();
+      requestPlayPauseFocus |= !shouldShowPlayButton && playButton.isFocused();
       requestPlayPauseAccessibilityFocus |=
           Util.SDK_INT < 21
               ? requestPlayPauseFocus
-              : (shouldShowPauseButton && Api21.isAccessibilityFocused(playButton));
-      playButton.setVisibility(shouldShowPauseButton ? GONE : VISIBLE);
+              : (!shouldShowPlayButton && Api21.isAccessibilityFocused(playButton));
+      playButton.setVisibility(shouldShowPlayButton ? VISIBLE : GONE);
     }
     if (pauseButton != null) {
-      requestPlayPauseFocus |= !shouldShowPauseButton && pauseButton.isFocused();
+      requestPlayPauseFocus |= shouldShowPlayButton && pauseButton.isFocused();
       requestPlayPauseAccessibilityFocus |=
           Util.SDK_INT < 21
               ? requestPlayPauseFocus
-              : (!shouldShowPauseButton && Api21.isAccessibilityFocused(pauseButton));
-      pauseButton.setVisibility(shouldShowPauseButton ? VISIBLE : GONE);
+              : (shouldShowPlayButton && Api21.isAccessibilityFocused(pauseButton));
+      pauseButton.setVisibility(shouldShowPlayButton ? GONE : VISIBLE);
     }
     if (requestPlayPauseFocus) {
       requestPlayPauseFocus();
@@ -1081,19 +1080,19 @@ public class LegacyPlayerControlView extends FrameLayout {
   }
 
   private void requestPlayPauseFocus() {
-    boolean shouldShowPauseButton = shouldShowPauseButton();
-    if (!shouldShowPauseButton && playButton != null) {
+    boolean shouldShowPlayButton = Util.shouldShowPlayButton(player);
+    if (shouldShowPlayButton && playButton != null) {
       playButton.requestFocus();
-    } else if (shouldShowPauseButton && pauseButton != null) {
+    } else if (!shouldShowPlayButton && pauseButton != null) {
       pauseButton.requestFocus();
     }
   }
 
   private void requestPlayPauseAccessibilityFocus() {
-    boolean shouldShowPauseButton = shouldShowPauseButton();
-    if (!shouldShowPauseButton && playButton != null) {
+    boolean shouldShowPlayButton = Util.shouldShowPlayButton(player);
+    if (shouldShowPlayButton && playButton != null) {
       playButton.sendAccessibilityEvent(AccessibilityEvent.TYPE_VIEW_FOCUSED);
-    } else if (shouldShowPauseButton && pauseButton != null) {
+    } else if (!shouldShowPlayButton && pauseButton != null) {
       pauseButton.sendAccessibilityEvent(AccessibilityEvent.TYPE_VIEW_FOCUSED);
     }
   }
@@ -1200,13 +1199,13 @@ public class LegacyPlayerControlView extends FrameLayout {
         switch (keyCode) {
           case KeyEvent.KEYCODE_MEDIA_PLAY_PAUSE:
           case KeyEvent.KEYCODE_HEADSETHOOK:
-            dispatchPlayPause(player);
+            Util.handlePlayPauseButtonAction(player);
             break;
           case KeyEvent.KEYCODE_MEDIA_PLAY:
-            dispatchPlay(player);
+            Util.handlePlayButtonAction(player);
             break;
           case KeyEvent.KEYCODE_MEDIA_PAUSE:
-            dispatchPause(player);
+            Util.handlePauseButtonAction(player);
             break;
           case KeyEvent.KEYCODE_MEDIA_NEXT:
             player.seekToNext();
@@ -1220,36 +1219,6 @@ public class LegacyPlayerControlView extends FrameLayout {
       }
     }
     return true;
-  }
-
-  private boolean shouldShowPauseButton() {
-    return player != null
-        && player.getPlaybackState() != Player.STATE_ENDED
-        && player.getPlaybackState() != Player.STATE_IDLE
-        && player.getPlayWhenReady();
-  }
-
-  private void dispatchPlayPause(Player player) {
-    @State int state = player.getPlaybackState();
-    if (state == Player.STATE_IDLE || state == Player.STATE_ENDED || !player.getPlayWhenReady()) {
-      dispatchPlay(player);
-    } else {
-      dispatchPause(player);
-    }
-  }
-
-  private void dispatchPlay(Player player) {
-    @State int state = player.getPlaybackState();
-    if (state == Player.STATE_IDLE) {
-      player.prepare();
-    } else if (state == Player.STATE_ENDED) {
-      seekTo(player, player.getCurrentMediaItemIndex(), C.TIME_UNSET);
-    }
-    player.play();
-  }
-
-  private void dispatchPause(Player player) {
-    player.pause();
   }
 
   @SuppressLint("InlinedApi")
@@ -1361,9 +1330,9 @@ public class LegacyPlayerControlView extends FrameLayout {
       } else if (rewindButton == view) {
         player.seekBack();
       } else if (playButton == view) {
-        dispatchPlay(player);
+        Util.handlePlayButtonAction(player);
       } else if (pauseButton == view) {
-        dispatchPause(player);
+        Util.handlePauseButtonAction(player);
       } else if (repeatToggleButton == view) {
         player.setRepeatMode(
             RepeatModeUtil.getNextRepeatMode(player.getRepeatMode(), repeatToggleModes));