+++ libraries/ui/src/main/java/androidx/media3/ui/PlayerControlView.java
@@ -19,11 +19,9 @@ import static androidx.media3.common.Player.COMMAND_GET_CURRENT_MEDIA_ITEM;
 import static androidx.media3.common.Player.COMMAND_GET_TIMELINE;
 import static androidx.media3.common.Player.COMMAND_GET_TRACKS;
 import static androidx.media3.common.Player.COMMAND_PLAY_PAUSE;
-import static androidx.media3.common.Player.COMMAND_PREPARE;
 import static androidx.media3.common.Player.COMMAND_SEEK_BACK;
 import static androidx.media3.common.Player.COMMAND_SEEK_FORWARD;
 import static androidx.media3.common.Player.COMMAND_SEEK_IN_CURRENT_MEDIA_ITEM;
-import static androidx.media3.common.Player.COMMAND_SEEK_TO_DEFAULT_POSITION;
 import static androidx.media3.common.Player.COMMAND_SEEK_TO_MEDIA_ITEM;
 import static androidx.media3.common.Player.COMMAND_SEEK_TO_NEXT;
 import static androidx.media3.common.Player.COMMAND_SEEK_TO_PREVIOUS;
@@ -75,7 +73,6 @@ import androidx.media3.common.Format;
 import androidx.media3.common.MediaLibraryInfo;
 import androidx.media3.common.Player;
 import androidx.media3.common.Player.Events;
-import androidx.media3.common.Player.State;
 import androidx.media3.common.Timeline;
 import androidx.media3.common.TrackGroup;
 import androidx.media3.common.TrackSelectionOverride;
@@ -978,17 +975,17 @@ public class PlayerControlView extends FrameLayout {
       return;
     }
     if (playPauseButton != null) {
-      boolean shouldShowPauseButton = shouldShowPauseButton();
+      boolean shouldShowPlayButton = Util.shouldShowPlayButton(player);
       @DrawableRes
       int drawableRes =
-          shouldShowPauseButton
-              ? R.drawable.exo_styled_controls_pause
-              : R.drawable.exo_styled_controls_play;
+          shouldShowPlayButton
+              ? R.drawable.exo_styled_controls_play
+              : R.drawable.exo_styled_controls_pause;
       @StringRes
       int stringRes =
-          shouldShowPauseButton
-              ? R.string.exo_controls_pause_description
-              : R.string.exo_controls_play_description;
+          shouldShowPlayButton
+              ? R.string.exo_controls_play_description
+              : R.string.exo_controls_pause_description;
       ((ImageView) playPauseButton)
           .setImageDrawable(getDrawable(getContext(), resources, drawableRes));
       playPauseButton.setContentDescription(resources.getString(stringRes));
@@ -1477,13 +1474,13 @@ public class PlayerControlView extends FrameLayout {
         switch (keyCode) {
           case KeyEvent.KEYCODE_MEDIA_PLAY_PAUSE:
           case KeyEvent.KEYCODE_HEADSETHOOK:
-            dispatchPlayPause(player);
+            Util.handlePlayPauseButtonAction(player);
             break;
           case KeyEvent.KEYCODE_MEDIA_PLAY:
-            dispatchPlay(player);
+            Util.handlePlayButtonAction(player);
             break;
           case KeyEvent.KEYCODE_MEDIA_PAUSE:
-            dispatchPause(player);
+            Util.handlePauseButtonAction(player);
             break;
           case KeyEvent.KEYCODE_MEDIA_NEXT:
             if (player.isCommandAvailable(COMMAND_SEEK_TO_NEXT)) {
@@ -1537,41 +1534,6 @@ public class PlayerControlView extends FrameLayout {
         && player.isCommandAvailable(COMMAND_PLAY_PAUSE)
         && (!player.isCommandAvailable(COMMAND_GET_TIMELINE)
             || !player.getCurrentTimeline().isEmpty());
-  }
-
-  private boolean shouldShowPauseButton() {
-    return player != null
-        && player.getPlaybackState() != Player.STATE_ENDED
-        && player.getPlaybackState() != Player.STATE_IDLE
-        && player.getPlayWhenReady();
-  }
-
-  private void dispatchPlayPause(Player player) {
-    @State int state = player.getPlaybackState();
-    if (state == Player.STATE_IDLE || state == Player.STATE_ENDED || !player.getPlayWhenReady()) {
-      dispatchPlay(player);
-    } else {
-      dispatchPause(player);
-    }
-  }
-
-  private void dispatchPlay(Player player) {
-    @State int state = player.getPlaybackState();
-    if (state == Player.STATE_IDLE && player.isCommandAvailable(COMMAND_PREPARE)) {
-      player.prepare();
-    } else if (state == Player.STATE_ENDED
-        && player.isCommandAvailable(COMMAND_SEEK_TO_DEFAULT_POSITION)) {
-      player.seekToDefaultPosition();
-    }
-    if (player.isCommandAvailable(COMMAND_PLAY_PAUSE)) {
-      player.play();
-    }
-  }
-
-  private void dispatchPause(Player player) {
-    if (player.isCommandAvailable(COMMAND_PLAY_PAUSE)) {
-      player.pause();
-    }
   }
 
   @SuppressLint("InlinedApi")
@@ -1743,7 +1705,7 @@ public class PlayerControlView extends FrameLayout {
           player.seekBack();
         }
       } else if (playPauseButton == view) {
-        dispatchPlayPause(player);
+        Util.handlePlayPauseButtonAction(player);
       } else if (repeatToggleButton == view) {
         if (player.isCommandAvailable(COMMAND_SET_REPEAT_MODE)) {
           player.setRepeatMode(