+++ libraries/ui/src/main/java/androidx/media3/ui/PlayerNotificationManager.java
@@ -18,11 +18,8 @@ package androidx.media3.ui;
 import static androidx.media3.common.Player.COMMAND_CHANGE_MEDIA_ITEMS;
 import static androidx.media3.common.Player.COMMAND_GET_CURRENT_MEDIA_ITEM;
 import static androidx.media3.common.Player.COMMAND_GET_TIMELINE;
-import static androidx.media3.common.Player.COMMAND_PLAY_PAUSE;
-import static androidx.media3.common.Player.COMMAND_PREPARE;
 import static androidx.media3.common.Player.COMMAND_SEEK_BACK;
 import static androidx.media3.common.Player.COMMAND_SEEK_FORWARD;
-import static androidx.media3.common.Player.COMMAND_SEEK_TO_DEFAULT_POSITION;
 import static androidx.media3.common.Player.COMMAND_SEEK_TO_NEXT;
 import static androidx.media3.common.Player.COMMAND_SEEK_TO_PREVIOUS;
 import static androidx.media3.common.Player.COMMAND_STOP;
@@ -1334,10 +1331,10 @@ public class PlayerNotificationManager {
       stringActions.add(ACTION_REWIND);
     }
     if (usePlayPauseActions) {
-      if (shouldShowPauseButton(player)) {
-        stringActions.add(ACTION_PAUSE);
-      } else {
+      if (Util.shouldShowPlayButton(player)) {
         stringActions.add(ACTION_PLAY);
+      } else {
+        stringActions.add(ACTION_PAUSE);
       }
     }
     if (useFastForwardAction && enableFastForward) {
@@ -1382,10 +1379,10 @@ public class PlayerNotificationManager {
     if (leftSideActionIndex != -1) {
       actionIndices[actionCounter++] = leftSideActionIndex;
     }
-    boolean shouldShowPauseButton = shouldShowPauseButton(player);
-    if (pauseActionIndex != -1 && shouldShowPauseButton) {
+    boolean shouldShowPlayButton = Util.shouldShowPlayButton(player);
+    if (pauseActionIndex != -1 && !shouldShowPlayButton) {
       actionIndices[actionCounter++] = pauseActionIndex;
-    } else if (playActionIndex != -1 && !shouldShowPauseButton) {
+    } else if (playActionIndex != -1 && shouldShowPlayButton) {
       actionIndices[actionCounter++] = playActionIndex;
     }
     if (rightSideActionIndex != -1) {
@@ -1398,12 +1395,6 @@ public class PlayerNotificationManager {
   protected boolean getOngoing(Player player) {
     int playbackState = player.getPlaybackState();
     return (playbackState == Player.STATE_BUFFERING || playbackState == Player.STATE_READY)
-        && player.getPlayWhenReady();
-  }
-
-  private boolean shouldShowPauseButton(Player player) {
-    return player.getPlaybackState() != Player.STATE_ENDED
-        && player.getPlaybackState() != Player.STATE_IDLE
         && player.getPlayWhenReady();
   }
 
@@ -1547,20 +1538,9 @@ public class PlayerNotificationManager {
       }
       String action = intent.getAction();
       if (ACTION_PLAY.equals(action)) {
-        if (player.getPlaybackState() == Player.STATE_IDLE
-            && player.isCommandAvailable(COMMAND_PREPARE)) {
-          player.prepare();
-        } else if (player.getPlaybackState() == Player.STATE_ENDED
-            && player.isCommandAvailable(COMMAND_SEEK_TO_DEFAULT_POSITION)) {
-          player.seekToDefaultPosition();
-        }
-        if (player.isCommandAvailable(COMMAND_PLAY_PAUSE)) {
-          player.play();
-        }
+        Util.handlePlayButtonAction(player);
       } else if (ACTION_PAUSE.equals(action)) {
-        if (player.isCommandAvailable(COMMAND_PLAY_PAUSE)) {
-          player.pause();
-        }
+        Util.handlePauseButtonAction(player);
       } else if (ACTION_PREVIOUS.equals(action)) {
         if (player.isCommandAvailable(COMMAND_SEEK_TO_PREVIOUS)) {
           player.seekToPrevious();